test: [move] test를 위한 내용들은 test 패키지로 이동
