+++ src/test/java/site/travellaboratory/be/user/application/_auth/PwInquiryServiceTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import site.travellaboratory.be.common.exception.BeApplicationException;
 import site.travellaboratory.be.common.exception.ErrorCodes;
-import site.travellaboratory.be.mock.user._auth.FakePwAnswerRepository;
-import site.travellaboratory.be.mock.user.FakeUserRepository;
+import site.travellaboratory.be.test.mock.user._auth.FakePwAnswerRepository;
+import site.travellaboratory.be.test.mock.user.FakeUserRepository;
 import site.travellaboratory.be.user.domain.User;
 import site.travellaboratory.be.user.domain._pw.PwAnswer;
 import site.travellaboratory.be.user.domain._pw.enums.PwAnswerStatus;
