+++ src/test/java/site/travellaboratory/be/user/application/_auth/UserRegistrationServiceTest.java
@@ -12,9 +12,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import site.travellaboratory.be.common.exception.BeApplicationException;
 import site.travellaboratory.be.common.exception.ErrorCodes;
-import site.travellaboratory.be.mock.user.FakeUserRepository;
-import site.travellaboratory.be.mock.user._auth.FakePwAnswerRepository;
-import site.travellaboratory.be.mock.user._auth.FakePwQuestionRepository;
+import site.travellaboratory.be.test.mock.user.FakeUserRepository;
+import site.travellaboratory.be.test.mock.user._auth.FakePwAnswerRepository;
+import site.travellaboratory.be.test.mock.user._auth.FakePwQuestionRepository;
 import site.travellaboratory.be.user.domain.User;
 import site.travellaboratory.be.user.domain._auth.request.UserJoinRequest;
 import site.travellaboratory.be.user.domain._pw.PwQuestion;
